kernel: do PROPOGATE_ERR grunt work to propogate errors and get a more accurate errno

+++ kernel/include/proc.h
@@ -110,7 +110,7 @@ exit(int) __attribute__((noreturn));
 pid_t
 fork(bool virtual);
 int
-growproc(int);
+growproc(intptr_t);
 int
 kill(pid_t, int);
 struct cpu *
